Split link.rs
